requirements.txt
PyQt6==6.9.1
pywin32==306
keyboard==0.13.5
Pillow==10.1.0
pyinstaller==6.3.0
